Complex internal rewrite with API consequences.
This extracted the node implementation to use dns-query instead of doh-query, as node's resolvers do not support ttl. This also equalizes --dns and --doh entries making the cli configuration much more streamlined.
In preparation for multiple entries this also changes the return statement to an array of links rather than a single entry causing the absence of conflict entries.

Closes #5
Closes #3